pyproject.toml
[project]
authors = [
    { name = "Song Feng", email = "song.feng@pnnl.gov" },
    { name = "Jordan Rozum", email = "jordan.rozum@pnnl.gov" },
]
description = "Add a short description here"
name = "proteometer"
requires-python = ">= 3.12"
version = "0.4.0"
dependencies = [
    "numpy>=2.2.1,<3",
    "pandas>=2.2.3,<3",
    "scipy>=1.14.1,<2",
    "statsmodels>=0.14.4,<0.15",
    "biopython>=1.84,<2",
    "matplotlib>=3.10.1,<4",
    "pingouin>=0.5.5,<0.6",
    "seaborn>=0.13.2,<0.14",
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-64", "osx-arm64", "linux-64", "win-64", "linux-aarch64"]

[tool.pixi.pypi-dependencies]
proteometer = { path = ".", editable = true }

[tool.pixi.tasks]
test = { cmd = "pytest tests" }
doctest = { cmd = "python -m doctest src/proteometer/*.py" }
lint = { cmd = "ruff check src tests" }
typecheck = { cmd = "pyright src tests" }
doc = { cmd = "sphinx-build -M html docs docs/_build" }
all_checks = { cmd = "ruff check src tests && pyright src tests && python -m doctest src/proteometer/*.py && pytest tests" }

[tool.pixi.dependencies]
python = ">=3.12.5,<3.13"

[dependency-groups]
dev = [
    "pandas-stubs",
    "scipy-stubs",
    "microsoft-python-type-stubs @ git+https://github.com/microsoft/python-type-stubs.git",
]
demo = [
    "matplotlib-venn>=1.1.2,<2",
    "ipykernel",
    "upsetplot>=0.9.0,<0.10",
    "gseapy>=1.1.11,<2",
]
test = ["pytest"]
lint = ["ruff", "pyright>=1.1.401"]
docs = [
    "sphinx<9.0.0",
    "sphinx-toolbox",
    "pydata-sphinx-theme",
    "sphinx-copybutton",
    "sphinx-autoapi",
]

[tool.pixi.environments]
default = { features = ["dev", "test", "lint", "docs"] }
demo = { features = ["dev", "test", "lint", "docs", "demo"] }
test = { features = ["dev", "test"] }
lint = { features = ["lint"] }
docs = { features = ["docs"] }

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = "src"

[tool.pyright]
exclude = ["docs", "PhenoMeters_proteometer.egg-info", ".pixi/**"]
include = ["src/proteometer"]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnnecessaryCast = "error"
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnknownMemberType = "none"
pythonVersion = "3.12"
useLibraryCodeForTypes = false
